+++ src/lib.rs
@@ -1,5 +1,6 @@
 mod background;
+pub mod client;
 mod errors;
 mod requests;
 mod responses;
-pub mod sync;
+mod websocket;
