+++ src/responses.rs
@@ -4,8 +4,6 @@ use serde::Deserialize;
 pub struct Response {
     pub id: String,
 
-    pub error: Option<String>,
-
     #[serde(flatten)]
     pub response: ResponseType,
 }
@@ -18,4 +16,6 @@ pub enum ResponseType {
     Publish { seq: u64 },
     #[serde(rename = "subscribe")]
     Subscribe(String),
+    #[serde(rename = "error")]
+    Error(String),
 }
